Complete lesson 06 reusable components and refactoring
