+++ src/components/TaskFilter.jsx
@@ -1,0 +1,10 @@
+export default function TaskFilter({ filter, setFilter }) {
+  return (
+    <div>
+      <button onClick={() => setFilter('all')}>All</button>
+      <button onClick={() => setFilter('completed')}>Completed</button>
+      <button onClick={() => setFilter('pending')}>Pending</button>
+      <p>Current filter: {filter}</p>
+    </div>
+  );
+}
