+++ src/components/TaskItem.jsx
@@ -1,0 +1,7 @@
+export default function TaskItem({ task }) {
+  return (
+    <li>
+      {task.title} {task.completed ? 'completed' : 'pending'}
+    </li>
+  );
+}
