+++ src/components/UserProfile.jsx
@@ -1,0 +1,3 @@
+export default function UserProfile() {
+  return <h2>Welcome, Student</h2>;
+}
