+++ src/utils/filterTasks.jsx
@@ -1,0 +1,11 @@
+export function filterTasks(tasks, filter) {
+  if (filter === 'completed') {
+    return tasks.filter((task) => task.completed);
+  }
+
+  if (filter === 'pending') {
+    return tasks.filter((task) => !task.completed);
+  }
+
+  return tasks;
+}
